Разделен на модули по заданию
